Reorganize e2e error handling tests into main AMM test
- Move error handling scenarios from separate file into amm-functionality.spec.ts
- Position error tests after normal AMM tests and before slippage test
- Consolidate test execution flow for better organization
